Replace ProductSwitcherEditionFooter with ProductSwitcherEditionMenuItem and enhance styling for menu items.
