+++ webapp/channels/src/components/global_header/left_controls/product_switcher_menu/product_switcher_edition_menuitem.test.tsx
@@ -11,7 +11,7 @@ import {TestHelper} from 'utils/test_helper';
 
 import type {GlobalState} from 'types/store';
 
-import ProductSwitcherEditionFooter from './product_switcher_edition_footer';
+import ProductSwitcherEditionFooter from './product_switcher_edition_menuitem';
 
 describe('ProductSwitcherEditionFooter', () => {
     const initialState: DeepPartial<GlobalState> = {
